requirements.txt
openai>=1.60.0
requests>=2.31.0
pypdf>=3.9.0
python-docx>=1.1.0
python-pptx>=0.6.21
flake8>=7.0.0
pylint>=3.0.0
bandit>=1.7.0
mypy>=1.10.0
